Crate accept

A library to create lexers or anything that does accepting easily. Inspired by Rob Pike’s lexer but made rusty.

Modules

combinators
Some structs that can be used to combined different things that can be accepted;

source

Macros

_and

_or

_seq

Structs

Any

Traits

Accept

Acceptor
An acceptor is something that can accept things from a Source. This trait is automatically implemented for all Sources.

Lexable
Something that can be lexed. This is implemented for most types already but can also be implemented for custom types, for example rowan SyntaxKind.