Trait accept::source::Source

pub trait Source {
    type Item;
    fn advance(&mut self) -> Option<Self::Item>;

    fn nth(&self, n: u32) -> Self::Item;

    fn nth_is_eof(&self, n: u32) -> bool;

    fn bump_raw(&mut self) -> Self::Item { ... }

    fn bump_expect(&mut self, msg: &str) -> Self::Item { ... }

    fn current(&self) -> Self::Item { ... }

    fn is_eof(&self) -> bool { ... }
}

A trait that defines the source that a lexer or parser can operate on.

Associated Types

type Item

Loading content...
Required methods

fn advance(&mut self) -> Option<Self::Item>

Advance the source.

fn nth(&self, n: u32) -> Self::Item

Get the nth item;

fn nth_is_eof(&self, n: u32) -> bool

Check is the nth item is eof.

Loading content...
Provided methods

fn bump_raw(&mut self) -> Self::Item

Bump the source. This advances the source but panics if there is nothing to be advanced.

fn bump_expect(&mut self, msg: &str) -> Self::Item

Bump the source. This advances the source but panics if there is nothing to be advanced. This will give a panic message.

fn current(&self) -> Self::Item

Get the current item. Equivalent to doing self.nth(0).

fn is_eof(&self) -> bool

Check if the current item is eof. Equivalent to doing self.nth_is_eof(0).

Loading content...