Trait accept::Lexable

pub trait Lexable<A: Source> {
    fn nth(self, p: &A, n: u32) -> bool
    where
        Self: Sized;

    fn peek(self, p: &A) -> bool
    where
        Self: Sized,
    { ... }
}

Something that can be lexed. This is implemented for most types already but can also be implemented for custom types, for example rowan SyntaxKind.

Required methods

fn nth(self, p: &A, n: u32) -> bool where
Self: Sized,

Loading content...
Provided methods

fn peek(self, p: &A) -> bool where
Self: Sized,

Loading content...